+++ private/buf/cmd/buf/command/mod/internal/internal.go
@@ -15,77 +15,14 @@
 package internal
 
 import (
-	"context"
 	"fmt"
 
-	"github.com/bufbuild/buf/private/buf/bufctl"
-	"github.com/bufbuild/buf/private/buf/bufworkspace"
 	"github.com/bufbuild/buf/private/bufpkg/bufcheck"
 	"github.com/bufbuild/buf/private/bufpkg/bufconfig"
-	"github.com/bufbuild/buf/private/bufpkg/bufmodule"
 	"github.com/bufbuild/buf/private/pkg/slicesext"
 	"github.com/bufbuild/buf/private/pkg/stringutil"
 	"github.com/spf13/pflag"
-	"go.uber.org/zap"
 )
-
-// Prune prunes the buf.lock.
-//
-// Used by both mod prune and mod update.
-func Prune(
-	ctx context.Context,
-	logger *zap.Logger,
-	controller bufctl.Controller,
-	workspaceDepManager bufworkspace.WorkspaceDepManager,
-	dirPath string,
-) error {
-	workspace, err := controller.GetWorkspace(ctx, dirPath, bufctl.WithIgnoreAndDisallowV1BufWorkYAMLs())
-	if err != nil {
-		return err
-	}
-	// Make sure the workspace builds.
-	if _, err := controller.GetImageForWorkspace(
-		ctx,
-		workspace,
-		bufctl.WithImageExcludeSourceInfo(true),
-	); err != nil {
-		return err
-	}
-	// Compute those dependencies that are in buf.yaml that are not used at all, and warn
-	// about them.
-	malformedDeps, err := bufworkspace.MalformedDepsForWorkspace(workspace)
-	if err != nil {
-		return err
-	}
-	for _, malformedDep := range malformedDeps {
-		switch t := malformedDep.Type(); t {
-		case bufworkspace.MalformedDepTypeUnused:
-			logger.Sugar().Warnf(
-				`Module %s is declared in your buf.yaml deps but is unused. This command only modifies buf.locks, not buf.yamls, please %s from your buf.yaml deps if it is not needed.`,
-				malformedDep.ModuleFullName(),
-				malformedDep.ModuleFullName(),
-			)
-		default:
-			return fmt.Errorf("unknown MalformedDepType: %v", t)
-		}
-	}
-	// Sep that actual computed remote dependencies based on imports. These are all
-	// that is needed for buf.lock.
-	depModules, err := bufmodule.RemoteDepsForModuleSet(workspace)
-	if err != nil {
-		return err
-	}
-	depModuleKeys, err := slicesext.MapError(
-		depModules,
-		func(remoteDep bufmodule.RemoteDep) (bufmodule.ModuleKey, error) {
-			return bufmodule.ModuleToModuleKey(remoteDep, workspaceDepManager.BufLockFileDigestType())
-		},
-	)
-	if err != nil {
-		return err
-	}
-	return workspaceDepManager.UpdateBufLockFile(ctx, depModuleKeys)
-}
 
 // BindLSRulesAll binds the all flag for an ls rules command.
 func BindLSRulesAll(flagSet *pflag.FlagSet, addr *bool, flagName string) {